+++ tests/input-output-pairs/attribute_expressions.spaces.gd
@@ -2,4 +2,6 @@ class X:
     func foo(xxxxxxxxxxx):
         var x = xxxxxxxxxxx.y.z
         var y = xxxxxxxxxxx.yyyyyyyyyyyy.zzzzzzzzzzz.qqqqqqqqqqqqq.wwwwwwwwww.eeeeeeeeeeee.rrrrrrrrrrrrr.tttttt
-        var z = xxxxxxxxxxx[[1,]].y.z.x[[1,]].c
+        var z = xxxxxxxxxxx[[
+            1,
+        ]].y.z.x[[1,]].c